refactor: FAQ tag labels are browser copy, not a shared contract
src/shared/faq/tags.ts carried FAQ_TAG_LABELS — the English display
strings 'Rules', 'Army List', 'Strategy' — beside the tag vocabulary
itself. The vocabulary (FAQ_TAG_VALUES) is a real cross-artifact
contract: convex/faq.ts, convex/lib/faqTags.ts and convex/migrations.ts
all parse against it. The labels are the browser's alone — four app
callers, zero server callers — so the Convex deploy was carrying UI copy
it has no use for.

They move to @ui/content/faqTagLabels.ts, beside the other data-to-words
modules in the kit (dates, assetPublishingStatus), keyed off the shared
FaqTag type so a new tag still forces a label. src/shared/faq/tags.ts is
left as pure vocabulary; src/shared/faq/validation.ts as pure schema —
which is what the shared layer is supposed to hold.
