+++ src/shared/faq/tags.ts
@@ -10,12 +10,3 @@ export const FAQ_TAG_VALUES = [
 export type FaqTag = (typeof FAQ_TAG_VALUES)[number];
 
 export const DEFAULT_FAQ_TAG: FaqTag = 'other';
-
-export const FAQ_TAG_LABELS: Record<FaqTag, string> = {
-  rules: 'Rules',
-  army_list: 'Army List',
-  strategy: 'Strategy',
-  balance: 'Balance',
-  errata: 'Errata',
-  other: 'Other',
-};
